engineapi/testing: refactor staticTxnProvider as package-local type
Move staticTxnProvider from txnprovider (public API) into the
testing_ namespace where it belongs. Replace Parameters.OverrideTxns
with Parameters.CustomProvider (txnprovider.TxnProvider interface)
so the builder is opaque to the provider implementation.

- builder.Parameters: OverrideTxns *[]types.Transaction → CustomProvider TxnProvider
- builder.Build: drop NewStaticTxnProvider call, just assign param.CustomProvider
- engineapi/testing_api: add local staticTxnProvider, create it after tx decode
- txnprovider/provider.go: remove NewStaticTxnProvider (reverted to main)
- txnprovider/shutter/pool_test.go: restore EmptyTxnProvider (reverted to main)
